+++ totalRP3/modules/register/main/register_tooltip.lua
@@ -105,6 +105,10 @@ local ANCHOR_TAB;
 local MATURE_CONTENT_ICON = Utils.str.texture("Interface\\AddOns\\totalRP3\\resources\\18_emoji.tga", 20);
 local registerTooltipModuleIsEnabled = false;
 
+local currentDate = date("*t");
+local seriousDay = currentDate.month == 4 and currentDate.day == 1
+local Rainbowify = TRP3_API.utils.Rainbowify;
+
 --*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 -- Config getters
 --*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
@@ -424,7 +428,11 @@ local function writeTooltipForCharacter(targetID, originalTexts, targetType)
 		completeName = crop(completeName, FIELDS_TO_CROP.NAME);
 	end
 
-	completeName = color:WrapTextInColorCode(completeName);
+	if seriousDay and getConfigValue("AF_STUFF") then
+		completeName = Rainbowify(completeName);
+	else
+		completeName = color:WrapTextInColorCode(completeName);
+	end
 
 	-- OOC
 	if info.character and info.character.RP ~= 1 then
@@ -598,7 +606,11 @@ local function writeTooltipForCharacter(targetID, originalTexts, targetType)
 				name = crop(name, FIELDS_TO_CROP.NAME);
 			end
 
-			name = color:WrapTextInColorCode(name);
+			if seriousDay and getConfigValue("AF_STUFF") then
+				name = Rainbowify(name);
+			else
+				name = color:WrapTextInColorCode(name);
+			end
 		end
 		tooltipBuilder:AddLine(loc.REG_TT_TARGET:format(name), 1, 1, 1, getSubLineFontSize());
 	end
@@ -735,6 +747,10 @@ local function writeCompanionTooltip(companionFullID, originalTexts, targetType,
 		petName = crop(petName, FIELDS_TO_CROP.NAME);
 	end
 
+	if seriousDay and getConfigValue("AF_STUFF") then
+		petName = Rainbowify(petName);
+	end
+
 
 	tooltipBuilder:AddLine(leftIcons .. "|cff" .. (info.NH or "ffffff") .. (petName or companionID), 1, 1, 1, getMainLineFontSize());
 
@@ -790,8 +806,12 @@ local function writeCompanionTooltip(companionFullID, originalTexts, targetType,
 			end
 		end
 
-		ownerFinalName = ownerColor:WrapTextInColorCode(ownerFinalName);
-		ownerFinalName = loc.REG_COMPANION_TF_OWNER:format(ownerFinalName);
+		if seriousDay and getConfigValue("AF_STUFF") then
+			ownerFinalName = Rainbowify(ownerFinalName);
+		else
+			ownerFinalName = ownerColor:WrapTextInColorCode(ownerFinalName);
+		end
+		ownerFinalName = loc("REG_COMPANION_TF_OWNER"):format(ownerFinalName);
 
 		tooltipBuilder:AddLine(ownerFinalName, 1, 1, 1, getSubLineFontSize());
 	end
@@ -888,6 +908,10 @@ local function writeTooltipForMount(ownerID, companionFullID, mountName)
 
 	if getConfigValue(CONFIG_CROP_TEXT) then
 		mountCustomName = crop(mountCustomName, FIELDS_TO_CROP.NAME);
+	end
+
+	if seriousDay and getConfigValue("AF_STUFF") then
+		mountCustomName = Rainbowify(mountCustomName);
 	end
 
 
@@ -1374,6 +1398,17 @@ local function onModuleInit()
 			},
 		}
 	}
+
+	if seriousDay then
+		registerConfigKey("AF_STUFF", true);
+		tinsert(CONFIG_STRUCTURE.elements, 2, {
+				inherit = "TRP3_ConfigCheck",
+				title = Rainbowify("Enable April Fools' joke"),
+				help = "Disable this option to remove this year's April Fools' joke.",
+				configKey = "AF_STUFF",
+			})
+	end
+
 	TRP3_API.configuration.registerConfigurationPage(CONFIG_STRUCTURE);
 end